Moves `KYCVerificationStatus` somewhere else.
